core: rework callable URL attributes
`url` was the only decorator in `sopel.module` to make use of `re`. Upon
removing its single call, I had to delete the import to appease flake8.

It struck me as odd a little while ago that `url` was the only decorator
adding *compiled* patterns. So now it doesn't. Compiling the patterns is
now left up to the bot to do when it loads the plugin and registers all
the callables.

As a bonus, this patch removes a gatekeeping check in the function that
dispatches URL callbacks. The check required any matching URL callback
to have the decorator-added attribute, as well, but it shouldn't be
necessary to check for that. Functions can only be returned by the
callback search if they've been registered, and to be registered they
must either already have that attribute OR be manually registered by the
containing plugin. Bugzilla does the latter, to support the fact that it
works based on a domain whitelist. That check would prevent it from
working at all.

The `url` decorator tests are also a bit simpler if the values it adds
are not compiled to Pattern objects, which is nice.