loader: use `inspect` to make callable docs consistent
Our own `trim_docstring` function is OK, but definitely not perfect.
What it returns can be inconsistently indented depending on whether the
docstring passed in is single- or multi-line, or whether a multi-line
docstring starts on line 0 or line 1 (next to or below the """).

I don't have any strong feelings about where to implement this. Maybe it
would make more sense to apply `inspect.cleandoc()` inside the logic of
`trim_docstring` instead of using `inspect.getdoc()` on the callable.
Either way, our help infrastructure will be easier to improve if we
eliminate anything like this that can cause inconsistent output.